Extracted main logic from the mixin.
