+++ src/main/java/tk/estecka/packrulemenus/PackRuleMenus.java
@@ -1,24 +1,22 @@
 package tk.estecka.packrulemenus;
 
-import java.util.Optional;
-import java.util.function.Consumer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.gui.screen.Screen;
-import net.minecraft.client.gui.screen.world.EditGameRulesScreen;
-import net.minecraft.world.GameRules;
-import tk.estecka.clothgamerules.api.ClothGamerulesScreenFactory;
+import net.minecraft.server.integrated.IntegratedServer;
+
 
 public class PackRuleMenus
 {
-	static public final Logger LOGGER = LoggerFactory.getLogger("Pack-Rule Menus");
+	static public final Logger LOGGER = LoggerFactory.getLogger("packrule-menus");
 
-	static public Screen CreateGameruleScreen(Screen parent, GameRules rules, Consumer<Optional<GameRules>> saveConsumer){
-		if (FabricLoader.getInstance().isModLoaded("cloth-gamerules"))
-			return ClothGamerulesScreenFactory.CreateScreen(parent, rules, saveConsumer);
-		else
-			return new EditGameRulesScreen(rules, saveConsumer.andThen( __->MinecraftClient.getInstance().setScreen(parent) ));
+	static public boolean CanModifyWorld(){
+		final MinecraftClient client = MinecraftClient.getInstance();
+		final IntegratedServer server = client.getServer();
+
+		return client.isIntegratedServerRunning()
+		    && server.getSaveProperties().areCommandsAllowed()
+		    && server.getOverworld() != null
+		    ;
 	}
 }
